+++ atlasdb-impl-shared/src/main/java/com/palantir/atlasdb/sweep/queue/SweepQueueCleaner.java
@@ -16,7 +16,15 @@
 
 package com.palantir.atlasdb.sweep.queue;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+import com.palantir.atlasdb.logging.LoggingArgs;
+import com.palantir.atlasdb.schema.generated.TargetedSweepTableFactory;
+import com.palantir.logsafe.SafeArg;
+
 public class SweepQueueCleaner {
+    private static final Logger log = LoggerFactory.getLogger(SweepQueueCleaner.class);
     private SweepableCells sweepableCells;
     private SweepableTimestamps sweepableTimestamps;
     private ShardProgress progress;
@@ -49,6 +57,9 @@ public class SweepQueueCleaner {
         if (minimumSweepPartitionNextIteration > lastSweptPartitionPreviously) {
             cleanDedicatedRows(shardStrategy, lastSweptPartitionPreviously);
             cleanNonDedicatedRow(shardStrategy, lastSweptPartitionPreviously);
+            log.info("Deleted persisted sweep queue information in table {} for partition {}.",
+                    LoggingArgs.tableRef(TargetedSweepTableFactory.of().getSweepableCellsTable(null).getTableRef()),
+                    SafeArg.of("partition", lastSweptPartitionPreviously));
         }
     }
 
@@ -68,14 +79,23 @@ public class SweepQueueCleaner {
         long minimumSweepPartitionNextIteration = SweepQueueUtils.tsPartitionCoarse(newProgress + 1);
         if (minimumSweepPartitionNextIteration > lastSweptPartitionPreviously) {
             sweepableTimestamps.deleteRow(shardStrategy, lastSweptPartitionPreviously);
+            log.info("Deleted persisted sweep queue information in table {} for partition {}.",
+                    LoggingArgs.tableRef(TargetedSweepTableFactory.of()
+                            .getSweepableTimestampsTable(null).getTableRef()),
+                    SafeArg.of("partition", lastSweptPartitionPreviously));
         }
     }
 
-    private void progressTo(ShardAndStrategy shardStrategy, long lastSweptTs) {
-        if (lastSweptTs < 0) {
+    private void progressTo(ShardAndStrategy shardStrategy, long newProgress) {
+        if (newProgress < 0) {
+            log.warn("Wasn't able to progress targeted sweep for {} since last swept timestamp {} is negative.",
+                    SafeArg.of("shardStrategy", shardStrategy.toText()), SafeArg.of("timestamp", newProgress));
             return;
         }
-        progress.updateLastSweptTimestamp(shardStrategy, lastSweptTs);
+        progress.updateLastSweptTimestamp(shardStrategy, newProgress);
+        log.info("Progressed last swept timestamp for {} to {}.",
+                SafeArg.of("shardStrategy", shardStrategy.toText()), SafeArg.of("timestamp", newProgress));
+
     }
 
     private boolean firstIterationOfSweep(long oldProgress) {